Fix C++20 build errors: WeakPointer spaceship operator, lambda captures, and GCC false positives
Rewrite WeakPointer comparison operators using C++20 spaceship operator,
replacing ~190 lines of non-member operators with 4 friend operators to
fix infinite recursion from C++20 reversed operator== candidates. Fix
deprecated implicit this capture via [=] in lambdas, GCC false positive
stringop-overflow in VectorProperty, GCC false positive restrict warning
in LegacyFileReaderService, and deprecated enum-float arithmetic in
ImageLiveWireContourModelFilter.

+++ Modules/Segmentation/Algorithms/mitkImageLiveWireContourModelFilter.cpp
@@ -335,7 +335,7 @@ void mitk::ImageLiveWireContourModelFilter::CreateDynamicCostMapByITK(
     // count pixel values
     // use scale factor to avoid mapping gradients between 0.0 and 1.0 to same bin
     histogram[static_cast<int>(gradientMagnImage->GetPixel((*pathIterator)) *
-                               ImageLiveWireContourModelFilter::CostFunctionType::MAPSCALEFACTOR)] += 1;
+                               static_cast<int>(ImageLiveWireContourModelFilter::CostFunctionType::MAPSCALEFACTOR))] += 1;
 
     pathIterator++;
   }
